// frontend/src/types/generated/api/models/StreamStatus.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 指示 YAESandBox.API.DTOs.WebSocket.DisplayUpdateDto 消息在流式传输过程中的状态。
 */
export enum StreamStatus {
    STREAMING = 'Streaming',
    COMPLETE = 'Complete',
    ERROR = 'Error',
}