// frontend/src/types/generated/api/models/UpdateMode.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 指示 YAESandBox.API.DTOs.WebSocket.DisplayUpdateDto.Content 的更新方式。
 */
export enum UpdateMode {
    FULL_SNAPSHOT = 'FullSnapshot',
    INCREMENTAL = 'Incremental',
}